+++ MarketPlace.Backend/MarketPlace/Microservices/User Service/Services.User.Tests/Integration/Tools/TestUserDatabaseFixture.cs
@@ -7,9 +7,13 @@ namespace UserService
     {
         public UserDbContext _context { get; private set; }
         private readonly IServiceScope _scope;
+        private bool _disposed;
 
         public TestUserDatabaseFixture()
         {
+            if(_disposed)
+                throw new ObjectDisposedException(ToString());
+
             var services = new ServiceCollection();
 
             services.AddDbContext<UserDbContext>(options =>
@@ -21,9 +25,13 @@ namespace UserService
         }
         public void Dispose()
         {
-            _context.Database.EnsureDeleted();
-            _context.Dispose();
-            _scope?.Dispose();
+            if (!_disposed)
+            {
+                _disposed = true;
+                _context.Database.EnsureDeleted();
+                _context.Dispose();
+                _scope?.Dispose();
+            }
         }
     }
 }
